requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
Flask==3.1.2
idna==3.11
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
paramiko==4.0.0
pycparser==2.23
PyNaCl==1.6.1
pyspnego==0.12.0
pywinrm==0.5.0
requests==2.32.5
requests_ntlm==1.3.0
urllib3==2.6.2
Werkzeug==3.1.4
xmltodict==1.0.2
gunicorn==21.2.0
curl_cffi==0.7.1
geoip2==4.8.0
PySocks==1.7.1
redis==5.0.1
rq==1.16.2
